Move Chapter related classes into chapter package
