+++ app/src/main/java/org/gdg/frisbee/android/activity/SearchActivity.java
@@ -22,6 +22,7 @@ import android.os.Bundle;
 
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
+import org.gdg.frisbee.android.chapter.MainActivity;
 
 import timber.log.Timber;
